chore: move enum to enum directory
