+++ src/main/java/com/group7/krisefikser/enums/Role.java
@@ -1,4 +1,4 @@
-package com.group7.krisefikser.model;
+package com.group7.krisefikser.enums;
 
 /**
  * Enum representing the different user roles in the system.
