+++ src/main/java/com/group7/krisefikser/utils/JwtUtils.java
@@ -7,7 +7,7 @@ import com.auth0.jwt.algorithms.Algorithm;
 import com.auth0.jwt.exceptions.JWTVerificationException;
 import com.auth0.jwt.interfaces.DecodedJWT;
 import com.group7.krisefikser.exception.JwtMissingPropertyException;
-import com.group7.krisefikser.model.Role;
+import com.group7.krisefikser.enums.Role;
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletResponse;
 import java.security.NoSuchAlgorithmException;
